Move test setup to separate module
